backend/requirements.txt
aiofiles==24.1.0
blinker==1.9.0
click==8.1.7
exceptiongroup==1.2.2
Flask==3.1.0
flask-marshmallow==1.2.1
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
Hypercorn==0.17.3
hyperframe==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
marshmallow==3.23.1
packaging==24.2
priority==2.0.0
psycopg2-binary==2.9.10
Quart==0.19.9
SQLAlchemy==2.0.36
taskgroup==0.0.0a4
toml==0.10.2
tomli==2.2.1
typing_extensions==4.12.2
Werkzeug==3.1.3
wsproto==1.2.0
python-dotenv
quart-flask-patch
SQLAlchemy-serializer
marshmallow-sqlalchemy